+++ src/alfasim_sdk/_internal/alfacase/case_description.py
@@ -762,6 +762,9 @@ class PumpEquipmentDescription:
     esp_parameters = attrib_enum(default=constants.EspParameters.UserDefined)
 
     esp_viscosity_model = attrib_enum(default=constants.PumpViscosityModel.NoModel)
+    esp_thermal_efficiency_model = attrib_enum(
+        default=constants.PumpThermalEfficiencyModel.Constant
+    )
 
 
 @attr.s(frozen=True, slots=True)